feat(structure): adjusting structure to organize files
