+++ src/components/Header/header.js
@@ -1,0 +1,19 @@
+import { Link } from "gatsby"
+import PropTypes from "prop-types"
+import React from "react"
+import { HeaderLaryout, NavLink } from "./style.js"
+
+const Header = () => (
+  <HeaderLaryout>
+    <div>
+        <NavLink to="/">
+          HOME
+        </NavLink>
+    </div>
+    <nav>
+      <Link>hello</Link>
+    </nav>
+  </HeaderLaryout>
+)
+
+export default Header
