+++ src/components/Header/style.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components'
+import { Link } from "gatsby"
+
+export const HeaderLaryout = styled.header`
+  background: transparent;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  padding: 1.5rem 6rem;
+`;
+
+export const NavLink = styled(Link)`
+  color: orange;
+`;
